modified:   src/touch/datastruct/datastruct.cpp 	modified:   src/touch/datastruct/datastruct.h 	modified:   src/touch/event/itspossibletoscroll.cpp

+++ src/touch/datastruct/datastruct.cpp
@@ -24,7 +24,7 @@ void datastruct::reorganize()
 
 }
 
-int datastruct::maxXIdOrizzonal()
+double datastruct::maxXIdOrizzonal()
 {
     int i, len = m_point.length();
